refactor: move logging into a folder and namespace
BREAKING CHANGE: Logging moved into a namespace,  use Mirror.Logging
