+++ Assets/Mirage/Authenticators/TimeoutAuthenticator.cs
@@ -1,6 +1,7 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using Mirage.Logging;
 
 namespace Mirage.Authenticators
 {
