+++ Assets/Mirage/Components/HeadlessFrameLimiter.cs
@@ -1,5 +1,6 @@
 using UnityEngine;
 using UnityEngine.Rendering;
+using Mirage.Logging;
 
 namespace Mirage
 {
