+++ Assets/Mirage/Components/NetworkSceneChecker.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.SceneManagement;
+using Mirage.Logging;
 
 namespace Mirage
 {
