+++ Assets/Mirage/Components/OnlineOfflineScene.cs
@@ -1,6 +1,7 @@
 using UnityEngine;
 using UnityEngine.SceneManagement;
 using UnityEngine.Serialization;
+using Mirage.Logging;
 
 namespace Mirage
 {
