+++ Assets/Mirage/Editor/LogLevelWindow.cs
@@ -2,6 +2,7 @@ using System.Collections.Generic;
 using System.Linq;
 using UnityEditor;
 using UnityEngine;
+using Mirage.Logging;
 
 namespace Mirage
 {
