+++ Assets/Mirage/Editor/NetworkBehaviourInspector.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.Reflection;
 using UnityEditor;
 using UnityEngine;
+using Mirage.Logging;
 
 namespace Mirage
 {
