+++ Assets/Mirage/Editor/NetworkMenu.cs
@@ -1,6 +1,7 @@
 using Mirage.KCP;
 using UnityEditor;
 using UnityEngine;
+using Mirage.Logging;
 
 namespace Mirage
 {
