+++ Assets/Mirage/Editor/NetworkScenePostProcess.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using UnityEditor;
 using UnityEditor.Callbacks;
 using UnityEngine;
+using Mirage.Logging;
 
 namespace Mirage
 {
