+++ Assets/Mirage/Editor/SceneDrawer.cs
@@ -1,5 +1,6 @@
 using UnityEditor;
 using UnityEngine;
+using Mirage.Logging;
 
 namespace Mirage
 {
