+++ Assets/Mirage/Runtime/CharacterSpawner.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.Serialization;
+using Mirage.Logging;
 
 namespace Mirage
 {
