+++ Assets/Mirage/Runtime/Logging.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a5b3d5184948c45ebb2b72410cf3098c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
