+++ Assets/Mirage/Runtime/NetworkReader.cs
@@ -11,6 +11,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Text;
 using UnityEngine;
+using Mirage.Logging;
 
 namespace Mirage
 {
