+++ Assets/Mirage/Runtime/NetworkServer.cs
@@ -5,6 +5,7 @@ using Cysharp.Threading.Tasks;
 using UnityEngine;
 using UnityEngine.Events;
 using UnityEngine.Serialization;
+using Mirage.Logging;
 
 namespace Mirage
 {
