+++ Assets/Mirage/Runtime/NetworkTime.cs
@@ -1,5 +1,6 @@
 using System;
 using UnityEngine;
+using Mirage.Logging;
 using Stopwatch = System.Diagnostics.Stopwatch;
 
 namespace Mirage
