+++ Assets/Mirage/Runtime/NetworkWriterPool.cs
@@ -1,5 +1,6 @@
 using System;
 using UnityEngine;
+using Mirage.Logging;
 
 namespace Mirage
 {
