+++ Assets/Mirage/Runtime/RemoteCallHelper.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using Cysharp.Threading.Tasks;
 using UnityEngine;
+using Mirage.Logging;
 
 namespace Mirage.RemoteCalls
 {
