+++ Assets/Mirage/Runtime/ServerObjectManager.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using Mirage.RemoteCalls;
 using UnityEngine;
 using UnityEngine.Serialization;
+using Mirage.Logging;
 
 namespace Mirage
 {
