+++ Assets/Mirage/Runtime/Transport/Kcp/KcpConnection.cs
@@ -6,6 +6,7 @@ using System.Net.Sockets;
 using System.Threading;
 using Cysharp.Threading.Tasks;
 using UnityEngine;
+using Mirage.Logging;
 
 namespace Mirage.KCP
 {
