+++ Assets/Mirage/Samples~/Chat/Scripts/ChatWindow.cs
@@ -2,6 +2,7 @@ using System.Collections;
 using UnityEngine;
 using UnityEngine.Serialization;
 using UnityEngine.UI;
+using Mirage.Logging;
 
 namespace Mirage.Examples.Chat
 {
