+++ Assets/Mirage/Samples~/MultipleAdditiveScenes/Scripts/Reward.cs
@@ -1,4 +1,5 @@
 using UnityEngine;
+using Mirage.Logging;
 
 namespace Mirage.Examples.MultipleAdditiveScenes
 {
