+++ Assets/Tests/Editor/LogFactoryTests.cs
@@ -1,6 +1,7 @@
 using NSubstitute;
 using NUnit.Framework;
 using UnityEngine;
+using Mirage.Logging;
 
 namespace Mirage
 {
